blast single column layout for smaller screen
