+++ src/content/app/tools/blast/components/blast-input-sequences/BlastInputSequences.scss
@@ -18,6 +18,14 @@
   background-color: $white;
 }
 
+.smallScreenHeader {
+  top: 49px;
+}
+
+.mainContainerSmall .header {
+  top: 49px;
+}
+
 .headerGroup {
   display: flex;
   align-items: center;
