+++ src/content/app/tools/blast/components/blast-input-sequences/BlastInputSequencesHeader.tsx
@@ -15,6 +15,7 @@
  */
 
 import React from 'react';
+import classNames from 'classnames';
 import { useSelector, useDispatch } from 'react-redux';
 
 import {
@@ -22,14 +23,10 @@ import {
   getUncommittedSequencePresence
 } from 'src/content/app/tools/blast/state/blast-form/blastFormSelectors';
 
-import {
-  updateEmptyInputDisplay,
-  switchToSpeciesStep
-} from 'src/content/app/tools/blast/state/blast-form/blastFormSlice';
+import { updateEmptyInputDisplay } from 'src/content/app/tools/blast/state/blast-form/blastFormSlice';
 
 import useBlastInputSequences from './useBlastInputSequences';
 
-import { SecondaryButton } from 'src/shared/components/button/Button';
 import PlusButton from 'src/shared/components/plus-button/PlusButton';
 import RadioGroup from 'src/shared/components/radio-group/RadioGroup';
 
@@ -61,10 +58,6 @@ const BlastInputSequencesHeader = (props: Props) => {
     setTimeout(() => scrollToLastInputBox(), 0);
   };
 
-  const onSwitchToSpecies = () => {
-    dispatch(switchToSpeciesStep());
-  };
-
   const scrollToLastInputBox = () => {
     const lastInputBox = document.querySelector(
       `.${sequenceBoxStyles.inputSequenceBox}:last-child`
@@ -81,8 +74,12 @@ const BlastInputSequencesHeader = (props: Props) => {
     ? sequences.length + 1
     : sequences.length;
 
+  const headerClass = classNames(styles.header, {
+    [styles.smallScreenHeader]: compact
+  });
+
   return (
-    <div className={styles.header}>
+    <div className={headerClass}>
       <div className={styles.headerGroup}>
         <span className={styles.headerTitle}>Sequences</span>
         <span className={styles.sequenceCounter}>{sequencesCount}</span>
@@ -100,14 +97,6 @@ const BlastInputSequencesHeader = (props: Props) => {
           onAdd={appendEmptyInput}
           disabled={!shouldEnableAddButton}
         />
-        {compact && (
-          <SecondaryButton
-            className={styles.blastAgainstButton}
-            onClick={onSwitchToSpecies}
-          >
-            Blast against
-          </SecondaryButton>
-        )}
       </div>
     </div>
   );
