+++ src/content/app/tools/blast/components/blast-species-selector/BlastSpeciesSelector.scss
@@ -13,6 +13,10 @@
   background-color: $white;
 }
 
+.smallScreenHeader {
+  top: 49px;
+}
+
 .headerGroup {
   display: flex;
   align-items: center;
