+++ src/content/app/tools/blast/components/blast-species-selector/BlastSpeciesSelectorHeader.tsx
@@ -15,15 +15,11 @@
  */
 
 import React from 'react';
+import classNames from 'classnames';
 import { useSelector, useDispatch } from 'react-redux';
 
-import {
-  switchToSequencesStep,
-  clearSelectedSpecies
-} from 'src/content/app/tools/blast/state/blast-form/blastFormSlice';
+import { clearSelectedSpecies } from 'src/content/app/tools/blast/state/blast-form/blastFormSlice';
 import { getSelectedSpeciesList } from 'src/content/app/tools/blast/state/blast-form/blastFormSelectors';
-
-import { SecondaryButton } from 'src/shared/components/button/Button';
 
 import styles from './BlastSpeciesSelector.scss';
 
@@ -37,43 +33,35 @@ const BlastSpeciesSelectorHeader = (props: Props) => {
 
   const selectedSpeciesList = useSelector(getSelectedSpeciesList);
 
-  const onSwitchToSequence = () => {
-    dispatch(switchToSequencesStep());
-  };
-
   const onClearAll = () => {
     dispatch(clearSelectedSpecies());
   };
 
+  const headerClass = classNames(styles.header, {
+    [styles.smallScreenHeader]: compact
+  });
+
   return (
-    <div className={styles.header}>
+    <div className={headerClass}>
       <div className={styles.headerGroup}>
         <span className={styles.headerTitle}>Blast against</span>
         <span className={styles.speciesCounter}>
           {selectedSpeciesList.length}
         </span>
         <span className={styles.maxSpecies}>of 7 species</span>
-        {compact && (
+      </div>
+      {compact && (
+        <span className={styles.clearAll} onClick={onClearAll}>
+          Clear all
+        </span>
+      )}
+      {!compact && (
+        <div className={styles.headerGroup}>
           <span className={styles.clearAll} onClick={onClearAll}>
             Clear all
           </span>
-        )}
-      </div>
-      <div className={styles.headerGroup}>
-        {!compact && (
-          <span className={styles.clearAll} onClick={onClearAll}>
-            Clear all
-          </span>
-        )}
-        {compact && (
-          <SecondaryButton
-            className={styles.sequencesButton}
-            onClick={onSwitchToSequence}
-          >
-            Sequences
-          </SecondaryButton>
-        )}
-      </div>
+        </div>
+      )}
     </div>
   );
 };
