+++ src/content/app/tools/shared/components/tools-top-bar/ToolsTopBar.scss
@@ -5,5 +5,5 @@
   background-color: $light-grey;
   padding: 24px 20px 20px 30px;
   box-shadow: 0 3px 5px $global-box-shadow;
-  margin-bottom: 21px;
+  margin-bottom: 20px;
 }
